:truck: EW-004: relocate queuelist to components
